transferred map loading functionality to redux
